+++ python/ray/rllib/test/test_filters.py
@@ -2,11 +2,10 @@ from __future__ import absolute_import
 from __future__ import division
 from __future__ import print_function
 
-import os
-import time
 import unittest
+import numpy as np
 
-from ray.rllib.optimizers import Evaluator
+from ray.rllib.utils.filter import RunningStat, MeanStdFilter
 
 
 class RunningStatTest(unittest.TestCase):
@@ -20,9 +19,9 @@ class RunningStatTest(unittest.TestCase):
                 li.append(val)
                 m = np.mean(li, axis=0)
                 self.assertTrue(np.allclose(rs.mean, m))
-                v = np.square(m) if (len(li) == 1) else np.var(li, ddof=1, axis=0)
+                v = (np.square(m) if (len(li) == 1)
+                     else np.var(li, ddof=1, axis=0))
                 self.assertTrue(np.allclose(rs.var, v))
-
 
     def testCombiningStat(self):
         for shape in [(), (3,), (3, 4)]:
@@ -33,7 +32,7 @@ class RunningStatTest(unittest.TestCase):
             for _ in range(5):
                 val = np.random.randn(*shape)
                 rs1.push(val)
-            rs.push(val)
+                rs.push(val)
                 li.append(val)
             for _ in range(9):
                 rs2.push(val)
@@ -43,18 +42,19 @@ class RunningStatTest(unittest.TestCase):
             assert np.allclose(rs.mean, rs1.mean)
             assert np.allclose(rs.std, rs1.std)
 
+
 class MSFTest(unittest.TestCase):
     def testBasic(self):
         """Test MeanStdFilter with scalar, vector, iamge"""
         # TODO(rliaw)
         msf = MeanStdFilter(())
-        for i in range(5)
+        for i in range(5):
             msf(1)
         msf = MeanStdFilter(())
-        for i in range(5)
+        for i in range(5):
             msf(1)
         msf = MeanStdFilter(())
-        for i in range(5)
+        for i in range(5):
             msf(1)
 
     def testMerge(self):
